other-repo-dir/requirements.txt
# Core dependencies (Windows-compatible versions)
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-huggingface>=0.0.3
chromadb>=0.4.0
sentence-transformers>=2.2.0
huggingface-hub>=0.16.0

# Document processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
markdown>=3.4.0
# unstructured>=0.10.0  # Commented out - can cause Windows issues

# LLM support (optional for Windows)
# llama-cpp-python>=0.2.0  # Commented out - can cause Windows build issues
transformers>=4.30.0
# torch>=2.0.0  # Commented out - large download, install separately if needed

# Utilities
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Testing and development (optional)
# pytest>=7.4.0  # Install separately if needed
# black>=23.0.0  # Install separately if needed  
# flake8>=6.0.0  # Install separately if needed

# Alternative minimal requirements for Windows
pathlib>=1.0.1
typing-extensions>=4.0.0
